metadata readers

+++ core/src/main/kotlin/ru/melonhell/umpa/core/utils/UmpaEulerAngle.kt
@@ -1,10 +1,12 @@
 package ru.melonhell.umpa.core.utils
 
-
+/**
+ * All values are in degrees
+ */
 data class UmpaEulerAngle(
-    val x: Double,
-    val y: Double,
-    val z: Double
+    val x: Float,
+    val y: Float,
+    val z: Float
 ) {
-    constructor() : this(0.0, 0.0, 0.0)
+    constructor() : this(0f, 0f, 0f)
 }
